test: 말 초기화 책임 변경 및 테스트 클래스 생성
